Remove react-native-youtube and allow the link to be handled natively
