--- ios/patches/XCDYouTubeVideoOperation.patch
@@ -1,11 +1,0 @@
---- Pods/XCDYouTubeKit/XCDYouTubeKit/XCDYouTubeVideoOperation.m	2021-05-28 20:53:40.000000000 -0400
-+++ XCDYouTubeVideoOperation.m	2021-05-28 20:55:18.000000000 -0400
-@@ -119,7 +119,7 @@
- 		NSString *eventLabel = [self.eventLabels objectAtIndex:0];
- 		[self.eventLabels removeObjectAtIndex:0];
- 		
--		NSDictionary *query = @{ @"video_id": self.videoIdentifier, @"hl": self.languageIdentifier, @"el": eventLabel, @"ps": @"default" };
-+        NSDictionary *query = @{ @"video_id": self.videoIdentifier, @"hl": self.languageIdentifier, @"el": eventLabel, @"ps": @"default", @"html5": @"1"};
- 		NSString *queryString = XCDQueryStringWithDictionary(query);
- 		NSURL *videoInfoURL = [NSURL URLWithString:[@"https://www.youtube.com/get_video_info?" stringByAppendingString:queryString]];
- 		[self startRequestWithURL:videoInfoURL type:XCDYouTubeRequestTypeGetVideoInfo];
