+++ package.json
@@ -83,7 +83,6 @@
     "react-native-vector-icons": "9.1.0",
     "react-native-video": "5.2.0",
     "react-native-webview": "11.18.1",
-    "react-native-youtube": "2.0.2",
     "react-syntax-highlighter": "15.5.0",
     "reanimated-bottom-sheet": "1.0.0-alpha.22",
     "rn-placeholder": "3.0.3",
